uninstall.phpの条件チェックを修正

+++ uninstall.php
@@ -1,6 +1,7 @@
 <?php
 
-if (!defined('ABSPATH') && !defined('WP_UNINSTALL_PLUGIN'))
-  exit();
+if ( ! defined( 'ABSPATH' ) || ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	exit;
+}
 
-delete_option('litoal-setting');
+delete_option( 'litoal-setting' );
